requirements.txt

fastapi==0.95.2
uvicorn==0.22.0
motor==3.1.1
python-dotenv==1.0.0
redis==4.5.5
pytest==7.4.0
pytest-asyncio==0.21.0
aiofiles==23.1.0
pymongo==4.5.0
bson
python-dotenv
boto3==1.28.40
aioboto3==11.3.0  # Асинхронная версия boto3 для работы с FastAPI